Day #6 Learning Laravel
Today i learned about how laravel manages sign ups, and it is actually cool that i somewhat learned how it works, i also learned about middleware being a good way to route all the routes into one single void function, named it Routes, i also didn't understand the @can of a tutorial so i just used a basic if function in the chirp (why? we want the user to only see the edit && delete button on his messages only).
And i also learned but Policies (not sure why they are helpful) but you can say they are allowed methods for the Controllers.

+++ app/Http/Controllers/ChirpController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chirp;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -38,10 +40,7 @@ public function store(Request $request)
         'message.max' => 'chirps must be 255 characters or less.',
     ]);
 
-    Chirp::create([
-        'message' => $validated['message'],
-        'user_id' => null,
-    ]);
+    auth()->user()->chirps()->create($validated); 
 
     return redirect('/')->with('success', 'Your chirp has been posted!');
 } 
@@ -59,7 +58,8 @@ public function store(Request $request)
      */
     public function edit(Chirp $chirp)
     {
-        
+        $this->authorize('update', $chirp);
+ 
         return view('chirps.edit', compact('chirp'));
     }
 
@@ -68,6 +68,8 @@ public function store(Request $request)
      */
     public function update(Request $request, Chirp $chirp)
     {
+
+        $this->authorize('update',$chirp);
          $validated = $request->validate([
             'message' => 'required|string|max:255',
             ], [
@@ -83,7 +85,11 @@ public function store(Request $request)
      */
     public function destroy(Chirp $chirp)
     {
+
+        $this->authorize('delete',$chirp);
+        //deletes the user;
         $chirp->delete();
+        //rerouting with success
         return redirect('/')->with('success', 'Chirp successfully deleted!');
     }
 }
